pyproject.toml
[tool.poetry]
name = "ptmpdy"
version = "0.1.0"
description = ""
authors = ["hayata-yamamoto <hayata.yamamoto.work@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
ptmpdy = "ptmpdy.main:app"

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^6.0"
typer = {extras = ["all"], version = "^0.7.0"}
rich = "^12.6.0"
Jinja2 = "^3.1.2"

[tool.poetry.dev-dependencies]
flake8 = "^6.0.0"
mypy = "^0.991"
black = "^22.12.0"
isort = "^5.11.4"
pytest = "^7.2.0"
autoflake = "^2.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
